+++ example/complete/controllers1/middleware_controllers/MiddlewareController.js
@@ -1,0 +1,7 @@
+module.exports = {
+  get,
+};
+
+function get(req, res) {
+  res.send(`Request with middleware : ${req.middleware_name}`);
+}