+++ example/complete/middlewares/withName/getName.js
@@ -1,0 +1,8 @@
+module.exports = {
+  getName,
+};
+
+function getName(req, res, next) {
+  req.middleware_name = 'getName';
+  next();
+}