+++ example/controllers/SubDirectory/SubSubDirectory/HomeController.js
@@ -1,9 +1,0 @@
-module.exports = {home, getParams};
-
-function home(req, res) {
-  res.send('Home');
-};
-
-function getParams(req, res) {
-  res.send({HomeController: req.params.id});
-}