+++ src/lib/routes.js
@@ -1,13 +1,11 @@
 const debug = require('debug')('ExpressImpRouter.routes');
 const routeExtractor = require('../extract/routes');
-const actionsGenerator = require('./actions-generator');
 
 let isDebug = false;
 let routes = [];
 
 module.exports = {
   extract,
-  generate,
   enableDebug,
   get,
 };
@@ -19,14 +17,23 @@ function get(debug) {
     return routes.filter(route => route.generated);
   } else {
     const colors = require('colors');
-    return routes.map((obj) => {
-      let color = colors.white, status = '';
+    colors.setTheme({
+      warn: 'yellow',
+      error: 'red'
+    });
+    let color = colors.white, status = '';
 
-      colors.setTheme({
-        warn: 'yellow',
-        error: 'red'
-      });
+    const debugRoutes = [];
+    let currentRoutes = [];
+    let currentConfigName = null;
 
+    routes
+    .sort((a, b) => {
+      const textA = a.config_name.toUpperCase();
+      const textB = b.config_name.toUpperCase();
+      return (textA < textB) ? -1 : (textA > textB) ? 1 : 0;
+    })
+    .map((obj) => {
       if(!obj.generated) {
         color = colors.red;
         status = 'Route not generated';
@@ -38,107 +45,29 @@ function get(debug) {
         }
       }
 
-      return {
+      if(!currentConfigName) {
+        currentConfigName = obj.config_name;
+      }
+
+      if(currentConfigName && currentConfigName !== obj.config_name) {
+        debugRoutes.push(currentRoutes);
+        currentRoutes = [];
+        currentConfigName = obj.config_name;
+      }
+
+      currentRoutes.push({
         status: color(status),
         method: color(obj.method.toUpperCase()),
         route: color(obj.route),
         controller: color(obj.debug.controller ? obj.debug.controller : 'N/A'),
         action: color(obj.debug.action ? obj.debug.action : 'N/A'),
         message: color(obj.debug.message ? obj.debug.message : 'N/A'),
-      };
+      });
     });
+
+    debugRoutes.push(currentRoutes);
+    return debugRoutes;
   }
-}
-
-/**
- * Generate routes
- */
-function generate() {
-  debug('Start generating routes');
-  for(let i in routes) {
-    generateRoute(routes[i]);
-  }
-
-  dedupe();
-
-  debug('Routes generating : done');
-}
-
-/**
- * Dedupe routes
- */
-function dedupe() {
-  debug('Start deduping');
-  const routeValidated = [];
-  let index = -1;
-
-  for(let i in routes) {
-    if(routes[i].static) {
-      continue;
-    }
-    let strToCompare = routes[i].method+routes[i].route;
-    let strToCompareHard = routes[i].method+routes[i].route.replace(/(\(.+\))/gi, '');
-
-    if((index = routeValidated.indexOf(strToCompare)) !== -1) {
-      routes[i].debug.message = `This route already exists : ${routes[i].method.toUpperCase()} ${routes[i].route}`;
-      routes[i].debug.level = 'error';
-      routes[i].generated = false;
-    } else {
-      if((index = routeValidated.indexOf(strToCompareHard)) !== -1) {
-        routes[i].debug.message = `Maybe there is similar route : ${routes[i].method.toUpperCase()} ${routes[i].route} :: ${routeValidated[index].replace(/([a-z]+)(\/.+)/, '$1').toUpperCase()} ${routeValidated[index].replace(/([a-z]+)(\/.+)/, '$2')}`;
-        routes[i].debug.level = 'warn';
-        routes[i].generated = true;
-      }
-
-      routeValidated.push(routes[i].method+routes[i].route);
-    }
-  }
-
-  debug('Deduping : done');
-}
-
-/**
- * @param route
- * @return {*}
- */
-function generateRoute(route) {
-  if(route.controller === '_ANON_') {
-    if(route.static) {
-      try {
-        route = actionsGenerator.generateStatic(route);
-        route.generated = true;
-        return route;
-      } catch(e) {
-        const message = `"${route.method.toUpperCase()} ${route.route}" is ignored. Reason : ${e.message}`;
-        if(!route.debug.message) {
-          route.debug.message = message;
-        }
-        debug(message);
-        if(isDebug) {
-          console.warn(message);
-        }
-        return route;
-      }
-    }
-    route.generated = true;
-    return route;
-  }
-
-  try {
-    route = actionsGenerator.generate(route);
-    route.generated = true;
-  } catch(e) {
-    const message = `"${route.method.toUpperCase()} ${route.route}" is ignored. Reason : ${e.message}`;
-    if(!route.debug.message) {
-      route.debug.message = message;
-    }
-    debug(message);
-    if(isDebug) {
-      console.warn(message);
-    }
-  }
-
-  return route;
 }
 
 /**
@@ -150,13 +79,25 @@ function extract(configAll) {
 
   let jsonRoutesConfig = null;
   for(let i in configAll) {
+    let scopeRoutes = [];
+
     if(typeof configAll[i].routes === 'object') {
       jsonRoutesConfig = configAll[i].routes;
     } else {
       jsonRoutesConfig = require(configAll[i].routes);
     }
 
-    routes = routes.concat(routeExtractor.route(configAll[i], jsonRoutesConfig, isDebug));
+    scopeRoutes = scopeRoutes.concat(routeExtractor.route(configAll[i], jsonRoutesConfig, isDebug));
+
+    if(configAll[i].config && configAll[i].config.route_mode === 'strict') {
+      scopeRoutes = dedupe(scopeRoutes);
+    }
+
+    for(let j in scopeRoutes) {
+      scopeRoutes[j].config_name = 'Config'+i;
+    }
+
+    routes = routes.concat(scopeRoutes);
   }
   debug('Extract routes : done');
   return routes;
@@ -167,4 +108,38 @@ function extract(configAll) {
  */
 function enableDebug() {
   isDebug = true;
+}
+
+/**
+ * Dedupe routes
+ */
+function dedupe(scopeRoutes) {
+  debug('Start deduping');
+  const routeValidated = [];
+  let index = -1;
+
+  for(let i in scopeRoutes) {
+    if(scopeRoutes[i].static) {
+      continue;
+    }
+    let strToCompare = scopeRoutes[i].method+scopeRoutes[i].route;
+    let strToCompareHard = scopeRoutes[i].method+scopeRoutes[i].route.replace(/(\(.+\))/gi, '');
+
+    if((index = routeValidated.indexOf(strToCompare)) !== -1) {
+      scopeRoutes[i].debug.message = `This route already exists : ${scopeRoutes[i].method.toUpperCase()} ${scopeRoutes[i].route}`;
+      scopeRoutes[i].debug.level = 'error';
+      scopeRoutes[i].isDedupe = true;
+    } else {
+      if((index = routeValidated.indexOf(strToCompareHard)) !== -1) {
+        scopeRoutes[i].debug.message = `Maybe there is similar route : ${scopeRoutes[i].method.toUpperCase()} ${scopeRoutes[i].route} :: ${routeValidated[index].replace(/([a-z]+)(\/.+)/, '$1').toUpperCase()} ${routeValidated[index].replace(/([a-z]+)(\/.+)/, '$2')}`;
+        scopeRoutes[i].debug.level = 'warn';
+        scopeRoutes[i].generated = true;
+      }
+
+      scopeRoutes[i].isDedupe = false;
+      routeValidated.push(scopeRoutes[i].method+scopeRoutes[i].route);
+    }
+  }
+  debug('Deduping : done');
+  return scopeRoutes;
 }